// src/lib/api/schemas/orderTag.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * ciclo
 * OpenAPI spec version: 0.1.0
 */
import type { OrderTagId } from './orderTagId.ts';
import type { OrderTagStatus } from './orderTagStatus.ts';

export interface OrderTag {
  /** @nullable */
  color?: string | null;
  created_at: string;
  display_name: string;
  id: OrderTagId;
  slug: string;
  status: OrderTagStatus;
}